+++ src/__tests__/worker/env.d.ts
@@ -4,5 +4,6 @@ declare namespace Cloudflare {
   interface Env {
     DB: D1Database
     TEST_ROOM: DurableObjectNamespace<import('./fixture').TestRoom>
+    OWNER_ROOM: DurableObjectNamespace<import('./fixture').OwnerRoom>
   }
 }